+++ python/lsst/ts/MTDome/__init__.py
@@ -1,8 +1,8 @@
 # This file is part of ts_MTDome.
 #
 # Developed for the Vera Rubin Observatory Telescope and Site Systems.
-# This product includes software developed by the Vera Rubin Observatory
-# Project (https://www.lsst.org).
+# This product includes software developed by the LSST Project
+# (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
 # for details of code ownership.
 #
@@ -19,14 +19,15 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+try:
+    from .version import *
+except ModuleNotFoundError:
+    __version__ = "?"
+
+from .config_schema import CONFIG_SCHEMA
 from .mtdome_csc import *
 from .llc_configuration_limits import *
 from .mock_controller import *
 from .mock_llc import *
 from .on_off import OnOff
 from .response_code import ResponseCode
-
-try:
-    from .version import *
-except ModuleNotFoundError:
-    __version__ = "?"